+++ docs/files_dup.js
@@ -1,4 +1,45 @@
 var files_dup =
 [
-    [ "Vitro-SDK-Lib", "dir_171f5be7832b2d0bfd1182f219a0da8b.html", "dir_171f5be7832b2d0bfd1182f219a0da8b" ]
+    [ "can_layer.h", "df/dce/can__layer_8h.html", null ],
+    [ "canbus.h", "d1/dd8/canbus_8h.html", "d1/dd8/canbus_8h" ],
+    [ "canbus_frame.h", "d9/d8c/canbus__frame_8h.html", null ],
+    [ "_CAN.h", "d7/de9/__CAN_8h.html", null ],
+    [ "canbus_protocol_consts.h", "dc/d8f/canbus__protocol__consts_8h.html", "dc/d8f/canbus__protocol__consts_8h" ],
+    [ "transport_layer.h", "d6/d7f/transport__layer_8h.html", null ],
+    [ "consts.h", "d8/db5/consts_8h.html", "d8/db5/consts_8h" ],
+    [ "upgrade_controller.h", "da/d8a/upgrade__controller_8h.html", "da/d8a/upgrade__controller_8h" ],
+    [ "node_controller.h", "d6/d9f/node__controller_8h.html", "d6/d9f/node__controller_8h" ],
+    [ "debug.h", "db/d16/debug_8h.html", "db/d16/debug_8h" ],
+    [ "modules.h", "d3/d61/modules_8h.html", null ],
+    [ "namespaces.h", "d6/d7b/namespaces_8h.html", null ],
+    [ "errors.h", "de/d45/errors_8h.html", "de/d45/errors_8h" ],
+    [ "environment_interface.h", "de/d4f/environment__interface_8h.html", "de/d4f/environment__interface_8h" ],
+    [ "flash_interface.h", "d6/d4d/flash__interface_8h.html", null ],
+    [ "iot-block.h", "dd/d88/iot-block_8h.html", "dd/d88/iot-block_8h" ],
+    [ "abstract_sensor_driver.h", "d7/d99/abstract__sensor__driver_8h.html", null ],
+    [ "abstract_sensor_interface.h", "df/d63/abstract__sensor__interface_8h.html", null ],
+    [ "measurement_api.h", "d4/d99/measurement__api_8h.html", "d4/d99/measurement__api_8h" ],
+    [ "io_expander.h", "d8/dc5/io__expander_8h.html", "d8/dc5/io__expander_8h" ],
+    [ "shard_edge.h", "dc/dce/shard__edge_8h.html", [
+      [ "vitroio::sdk::ShardEdge", "d9/d30/classvitroio_1_1sdk_1_1ShardEdge.html", "d9/d30/classvitroio_1_1sdk_1_1ShardEdge" ]
+    ] ],
+    [ "shard-sdk.h", "d4/d3e/shard-sdk_8h.html", null ],
+    [ "timestamp.h", "d2/d02/timestamp_8h.html", [
+      [ "vitroio::sdk::impl::Timestamp::CurrentTime_t", "d2/db8/structvitroio_1_1sdk_1_1impl_1_1Timestamp_1_1CurrentTime__t.html", "d2/db8/structvitroio_1_1sdk_1_1impl_1_1Timestamp_1_1CurrentTime__t" ]
+    ] ],
+    [ "types.h", "d9/d49/types_8h.html", "d9/d49/types_8h" ],
+    [ "buffer.h", "d5/d08/buffer_8h.html", [
+      [ "vitroio::sdk::impl::Buffer", "db/d9a/classvitroio_1_1sdk_1_1impl_1_1Buffer.html", "db/d9a/classvitroio_1_1sdk_1_1impl_1_1Buffer" ]
+    ] ],
+    [ "list.h", "da/d23/list_8h.html", [
+      [ "vitroio::sdk::impl::ListNode< T >", "d1/da3/structvitroio_1_1sdk_1_1impl_1_1ListNode.html", "d1/da3/structvitroio_1_1sdk_1_1impl_1_1ListNode" ],
+      [ "vitroio::sdk::impl::List< T >", "d6/d16/classvitroio_1_1sdk_1_1impl_1_1List.html", "d6/d16/classvitroio_1_1sdk_1_1impl_1_1List" ]
+    ] ],
+    [ "scoped_ptr.h", "d0/d44/scoped__ptr_8h.html", [
+      [ "vitroio::sdk::impl::ScopedPtr< T >", "d0/d0a/classvitroio_1_1sdk_1_1impl_1_1ScopedPtr.html", "d0/d0a/classvitroio_1_1sdk_1_1impl_1_1ScopedPtr" ]
+    ] ],
+    [ "version.h", "df/dfc/version_8h.html", "df/dfc/version_8h" ],
+    [ "watchdog.h", "d9/da3/watchdog_8h.html", [
+      [ "vitroio::sdk::Watchdog", "d6/d81/classvitroio_1_1sdk_1_1Watchdog.html", "d6/d81/classvitroio_1_1sdk_1_1Watchdog" ]
+    ] ]
 ];