+++ docs/search/functions_11.js
@@ -1,5 +1,6 @@
 var searchData=
 [
-  ['updater_5frunner_0',['updater_runner',['../da/d87/classvitroio_1_1sdk_1_1impl_1_1Timestamp.html#aa9dc796b9df646d5a12af82d281c26f0',1,'vitroio::sdk::impl::Timestamp']]],
-  ['upgradecontroller_1',['UpgradeController',['../db/d78/classvitroio_1_1sdk_1_1impl_1_1UpgradeController.html#afcaf7a35ba3309168cf7646cc1cb693f',1,'vitroio::sdk::impl::UpgradeController']]]
+  ['updatenodeid_0',['updateNodeId',['../d6/dd0/classvitroio_1_1sdk_1_1Transport__layer.html#a68e22e6a0f161a1ccbf49d83b9364726',1,'vitroio::sdk::Transport_layer']]],
+  ['updater_5frunner_1',['updater_runner',['../da/d87/classvitroio_1_1sdk_1_1impl_1_1Timestamp.html#aa9dc796b9df646d5a12af82d281c26f0',1,'vitroio::sdk::impl::Timestamp']]],
+  ['upgradecontroller_2',['UpgradeController',['../db/d78/classvitroio_1_1sdk_1_1impl_1_1UpgradeController.html#afcaf7a35ba3309168cf7646cc1cb693f',1,'vitroio::sdk::impl::UpgradeController']]]
 ];